pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devflow-lab"
version = "0.1.0"
description = "Claude Skills + OpenSpec + GitHub PR automation — bundled open-source app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "anthropic>=0.28.0",
    "langgraph>=0.1.0",
    "langfuse>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
dev       = "uvicorn app.backend.main:app --reload --port 8000"
test      = "pytest tests/ -v"
lint      = "ruff check . && ruff format --check ."
fmt       = "ruff format ."
benchmark = "python -m app.backend.agentbench.benchmark"
eval      = "python scripts/run_eval.py"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"
